requirements.txt
anyio==3.6.1
asgiref==3.5.1
attrs==21.4.0
bcrypt==3.2.2
black==22.3.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.2.1
emails==0.6
fastapi==0.78.0
graypy==2.1.0
h11==0.13.0
hubspot-api-client==5.0.1
idna==3.3
iniconfig==1.1.1
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
Naked==0.1.31
packaging==21.3
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pycrypto==2.6.1
pydantic==1.9.0
pyparsing==3.0.9
pytest==7.1.2
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.27.1
rsa==4.8
shellescape==3.8.1
six==1.16.0
slowapi==0.1.5
sniffio==1.2.0
SQLAlchemy==1.4.36
SQLAlchemy-Utils==0.38.2
starlette==0.19.1
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
